[refactor] modify file name of bot_model.js and move it to views
