+++ backend/src/views/bot_reply_template.js
